iptu_lavras/requirements.txt
playwright
sqlalchemy
psycopg2-binary
python-dotenv
requests
SpeechRecognition
pydub